Added vehicle animation
critter vehicles are now functional. A buffered image was added to the
critter class with a line in the constructor of each critter type to
load the corresponding sprite sheet. Code was added in display critter
class to determine the direction of travel and snip out the
corresponding image from the sprite sheet.